Made Changes
Player Throw is fixed completely and he now throws correctly every time.
The doors can now take more than one input from multiple buttons.
Made it so that the E doesn't display when it shouldn't on the fire_bush, water_bush and eating_food.
Added fire idle to the audiocontroller.
Made it so that the Hud now has a selection sound.